Move gcs functions file in util folder.
